Version 3.26.20 (based on bleeding_edge revision r20866)
ES6: Add support for Map/Set forEach (Chromium issues 1793, 2323).

Performance and stability improvements on all platforms.

git-svn-id: http://v8.googlecode.com/svn/trunk@20867 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/x64/macro-assembler-x64.h
@@ -1465,7 +1465,7 @@ class MacroAssembler: public Assembler {
   // modified. It may be the "smi 1 constant" register.
   Register GetSmiConstant(Smi* value);
 
-  intptr_t RootRegisterDelta(ExternalReference other);
+  int64_t RootRegisterDelta(ExternalReference other);
 
   // Moves the smi value to the destination register.
   void LoadSmiConstant(Register dst, Smi* value);
